refactor: move container/iterator to their own files
