+++ src/core/geometry/octree.cpp
@@ -13,8 +13,8 @@
 #include <Eigen/Dense>
 
 #include "nuri/eigen_config.h"
+#include "nuri/core/container/container_ext.h"
 #include "nuri/core/geometry.h"
-#include "nuri/utils.h"
 
 namespace nuri {
 namespace {
