+++ src/fmt/pdb.cpp
@@ -36,6 +36,7 @@
 
 #include "nuri/eigen_config.h"
 #include "fmt_internal.h"
+#include "nuri/core/container/compact_map.h"
 #include "nuri/core/container/property_map.h"
 #include "nuri/core/element.h"
 #include "nuri/core/graph/graph.h"
